Acf {forecast}	R Documentation

(Partial) Autocorrelation Function Estimation

Description

The function Acf computes (and by default plots) an estimate of the autocorrelation function of a univariate time series. Function Pacf computes (and by default plots) an estimate of the partial autocorrelation function of a univariate time series. These improve the acf and pacf functions when applied to univariate time series. The main differences are that Acf does not plot a spike at lag 0 (which is redundant) and the horizontal axes show lags in time units rather than seasonal units.

Usage

Acf(x, lag.max=NULL, type=c("correlation", "partial"), 
   plot=TRUE, main=NULL, ylim=NULL, na.action=na.contiguous, ...)
Pacf(x, main=NULL, ...) 

Arguments

x
a univariate time series

lag.max
maximum lag at which to calculate the acf. Default is 10*log10(N/m) where N is the number of observations and m the number of series. Will be automatically limited to one less than the number of observations in the series.

type
character string giving the type of acf to be computed. Allowed values are "correlation" (the default) or "partial".

plot
logical. If TRUE (the default) the acf is plotted.

main
Title for plot

ylim
The y limits of the plot

na.action
function to handle missing values. Default is na.contiguous. Useful alternatives are na.pass and na.interp.

...
Additional arguments passed to acf.

Details

See the acf function in the stats package.

Value

See the acf function in the stats package.

Author(s)

Rob J Hyndman

See Also

acf, pacf, tsdisplay

Examples

Acf(wineind)
Pacf(wineind)

[Package forecast version 5.9 Index]