arima.errors {forecast}	R Documentation

ARIMA errors

Description

Returns original time series after adjusting for regression variables. These are not the same as the residuals. If there are no regression variables in the ARIMA model, then the errors will be identical to the original series. If there are regression variables in the ARIMA model, then the errors will be equal to the original series minus the effect of the regression variables, but leaving in the serial correlation that is modelled with the AR and MA terms. If you want the "residuals", then use residuals(z)..

Usage

arima.errors(z)

Arguments

z
Fitted ARIMA model from arima

Value

A time series containing the "errors".

Author(s)

Rob J Hyndman

See Also

arima, residuals

Examples

www.fit <- auto.arima(WWWusage)
www.errors <- arima.errors(www.fit)
par(mfrow=c(2,1))
plot(WWWusage)
plot(www.errors)

[Package forecast version 5.9 Index]