baggedETS {forecast}	R Documentation

Forecasting using the bagged ETS method

Description

The bagged ETS forecasting method.

Usage

baggedETS(y, bootstrapped_series = bld.mbb.bootstrap(y, 100), ...)

Arguments

y
A numeric vector or time series of class ts.

bootstrapped_series
bootstrapped versions of y.

...
Other arguments passed to ets.

Details

This function implements the bagged ETS forecasting method described in Bergmeir et al. The ets function is applied to all bootstrapped series. Using the default parameters, the function bld.mbb.bootstrap is used to calculate the bootstrapped series with the Box-Cox and Loess-based decomposition (BLD) bootstrap. The function forecast.baggedETS can then be used to calculate forecasts.

Value

Returns an object of class "baggedETS".

The function print is used to obtain and print a summary of the results.

models
A list containing the fitted ETS ensemble models.

method
The name of the forecasting method as a character string

y
The original time series.

bootstrapped_series
The bootstrapped series.

etsargs
The arguments passed through to ets.

fitted
Fitted values (one-step forecasts). The mean is of the fitted values is calculated over the ensemble.

residuals
Original values minus fitted values.

Author(s)

Christoph Bergmeir, Fotios Petropoulos

References

Bergmeir, C., R. J. Hyndman, and J. M. Benitez (2016). Bagging Exponential Smoothing Methods using STL Decomposition and Box-Cox Transformation. International Journal of Forecasting 32, 303-312.

Examples

fit <- baggedETS(WWWusage)
fcast <- forecast(fit)
plot(fcast)

[Package forecast version 8.1 Index]