bld.mbb.bootstrap {forecast}	R Documentation

Box-Cox and Loess-based decomposition bootstrap.

Description

Generates bootstrapped versions of a time series using the Box-Cox and Loess-based decomposition bootstrap.

Usage

bld.mbb.bootstrap(x, num, block_size = NULL)

Arguments

x
Original time series.

num
Number of bootstrapped versions to generate.

block_size
Block size for the moving block bootstrap.

Details

The procedure is described in Bergmeir et al. Box-Cox decomposition is applied, together with STL or Loess (for non-seasonal time series), and the remainder is bootstrapped using a moving block bootstrap.

Value

A list with bootstrapped versions of the series. The first series in the list is the original series.

Author(s)

Christoph Bergmeir, Fotios Petropoulos

References

Bergmeir, C., R. J. Hyndman, and J. M. Benitez (2016). Bagging Exponential Smoothing Methods using STL Decomposition and Box-Cox Transformation. International Journal of Forecasting 32, 303-312.

See Also

baggedETS.

Examples

bootstrapped_series <- bld.mbb.bootstrap(WWWusage, 100)

[Package forecast version 8.12 Index]