easter {forecast}	R Documentation

Easter holidays in each season

Description

Returns a vector of 0's and 1's or fractional results if Easter spans March and April in the observed time period. Easter is defined as the days from Good Friday to Easter Sunday inclusively, plus optionally Easter Monday if easter.mon=TRUE.

Usage

easter(x, easter.mon = FALSE)

Arguments

x
Monthly or quarterly time series

easter.mon
If TRUE, the length of Easter holidays includes Easter Monday.

Details

Useful for adjusting calendar effects.

Value

Time series

Author(s)

Earo Wang

Examples

easter(wineind, easter.mon = TRUE)

[Package forecast version 8.12 Index]