forecast-package {forecast}	R Documentation

Forecasting Functions for Time Series and Linear Models

Description

Methods and tools for displaying and analysing univariate time series forecasts including exponential smoothing via state space models and automatic ARIMA modelling.

Details

Package: 	 forecast

Type: 	 Package

License: 	 GPL3

LazyLoad: 	 yes

Author(s)

Rob J Hyndman

Maintainer: Rob.Hyndman@monash.edu

[Package forecast version 8.12 Index]