forecast.baggedETS {forecast}	R Documentation

Forecasting using the bagged ETS method

Description

Returns forecasts and other information for bagged ETS models.

Usage

## S3 method for class 'baggedETS'
forecast(object, h = ifelse(frequency(object$x) > 1, 2 *
  frequency(object$x), 10), ...)

Arguments

object
An object of class "baggedETS" resulting from a call to baggedETS.

h
Number of periods for forecasting.

...
Additional arguments passed to ets

Details

Intervals are calculated as min and max values over the point forecasts from the ETS models in the ensemble. I.e., the intervals are not prediction intervals, but give an indication of how different the forecasts within the ensemble are.

Value

An object of class "forecast".

The function summary is used to obtain and print a summary of the results, while the function plot produces a plot of the forecasts and prediction intervals.

An object of class "forecast" is a list containing at least the following elements:

model
A list containing information about the fitted model

method
The name of the forecasting method as a character string

mean
Point forecasts as a time series

lower
Lower limits for prediction intervals

upper
Upper limits for prediction intervals

level
The confidence values associated with the prediction intervals

x
The original time series (either object itself or the time series used to create the model stored as object).

xreg
The external regressors used in fitting (if given).

residuals
Residuals from the fitted model. That is x minus fitted values.

fitted
Fitted values (one-step forecasts)

...
Other arguments

Author(s)

Christoph Bergmeir, Fotios Petropoulos

References

Bergmeir, C., R. J. Hyndman, and J. M. Benitez (2016). Bagging Exponential Smoothing Methods using STL Decomposition and Box-Cox Transformation. International Journal of Forecasting 32, 303-312.

See Also

baggedETS.

Examples

fit <- baggedETS(WWWusage)
fcast <- forecast(fit)
plot(fcast)

[Package forecast version 8.1 Index]