forecast.ets {forecast}	R Documentation

Forecasting using ETS models

Description

Returns forecasts and other information for univariate ETS models.

Usage

## S3 method for class 'ets'
forecast(object, h = ifelse(object$m > 1, 2 * object$m,
  10), level = c(80, 95), fan = FALSE, simulate = FALSE,
  bootstrap = FALSE, npaths = 5000, PI = TRUE,
  lambda = object$lambda, biasadj = NULL, ...)

Arguments

object
An object of class "ets". Usually the result of a call to ets.

h
Number of periods for forecasting

level
Confidence level for prediction intervals.

fan
If TRUE, level is set to seq(51,99,by=3). This is suitable for fan plots.

simulate
If TRUE, prediction intervals are produced by simulation rather than using analytic formulae. Errors are assumed to be normally distributed.

bootstrap
If TRUE, then prediction intervals are produced by simulation using resampled errors (rather than normally distributed errors).

npaths
Number of sample paths used in computing simulated prediction intervals.

PI
If TRUE, prediction intervals are produced, otherwise only point forecasts are calculated. If PI is FALSE, then level, fan, simulate, bootstrap and npaths are all ignored.

lambda
Box-Cox transformation parameter. If lambda="auto", then a transformation is automatically selected using BoxCox.lambda. The transformation is ignored if NULL. Otherwise, data transformed before model is estimated.

biasadj
Use adjusted back-transformed mean for Box-Cox transformations. If transformed data is used to produce forecasts and fitted values, a regular back transformation will result in median forecasts. If biasadj is TRUE, an adjustment will be made to produce mean forecasts and fitted values.

...
Other arguments.

Value

An object of class "forecast".

The function summary is used to obtain and print a summary of the results, while the function plot produces a plot of the forecasts and prediction intervals.

The generic accessor functions fitted.values and residuals extract useful features of the value returned by forecast.ets.

An object of class "forecast" is a list containing at least the following elements:

model
A list containing information about the fitted model

method
The name of the forecasting method as a character string

mean
Point forecasts as a time series

lower
Lower limits for prediction intervals

upper
Upper limits for prediction intervals

level
The confidence values associated with the prediction intervals

x
The original time series (either object itself or the time series used to create the model stored as object).

residuals
Residuals from the fitted model. For models with additive errors, the residuals are x - fitted values. For models with multiplicative errors, the residuals are equal to x /(fitted values) - 1.

fitted
Fitted values (one-step forecasts)

Author(s)

Rob J Hyndman

See Also

ets, ses, holt, hw.

Examples

fit <- ets(USAccDeaths)
plot(forecast(fit,h=48))

[Package forecast version 8.5 Index]