plot.forecast {forecast}	R Documentation

Forecast plot

Description

Plots historical data with forecasts and prediction intervals.

Usage

## S3 method for class 'forecast'
plot(x, include, plot.conf=TRUE, shaded=TRUE,
    shadebars=(length(x$mean)<5), shadecols=NULL, col=1, fcol=4,
    pi.col=1, pi.lty=2, ylim=NULL, main=NULL, ylab="", xlab="", type="l",
    flty=1, flwd=2, ...)

## S3 method for class 'splineforecast'
plot(x, fitcol=2, type="o", pch=19, ...)

Arguments

x
Forecast object produced by forecast.

include
number of values from time series to include in plot

plot.conf
Logical flag indicating whether to plot prediction intervals.

shaded
Logical flag indicating whether prediction intervals should be shaded (TRUE) or lines (FALSE)

shadebars
Logical flag indicating if prediction intervals should be plotted as shaded bars (if TRUE) or a shaded polygon (if FALSE). Ignored if shaded=FALSE. Bars are plotted by default if there are fewer than five forecast horizons.

shadecols
Colors for shaded prediction intervals. To get default colors used prior to v3.26, set shadecols="oldstyle".

col
Colour for the data line.

fcol
Colour for the forecast line.

flty
Line type for the forecast line.

flwd
Line width for the forecast line.

pi.col
If shade=FALSE and plot.conf=TRUE, the prediction intervals are plotted in this colour.

pi.lty
If shade=FALSE and plot.conf=TRUE, the prediction intervals are plotted using this line type.

ylim
Limits on y-axis

main
Main title

ylab
Y-axis label

xlab
X-axis label

fitcol
Line colour for fitted values.

type
1-character string giving the type of plot desired. As for plot.default.

pch
Plotting character (if type=="p" or type=="o").

...
additional arguments to plot.

Value

None.

Author(s)

Rob J Hyndman

References

Hyndman and Athanasopoulos (2014) Forecasting: principles and practice, OTexts: Melbourne, Australia. http://www.otexts.org/fpp/

See Also

plot.ts

Examples

deaths.fit <- hw(USAccDeaths,h=48)
plot(deaths.fit)

[Package forecast version 5.9 Index]