getResponse {forecast}	R Documentation

Get response variable from time series model.

Description

getResponse is a generic function for extracting the historical data from a time series model (including Arima, ets, ar, fracdiff), a linear model of class lm, or a forecast object. The function invokes particular methods which depend on the class of the first argument.

Usage

getResponse(object,...)

Arguments

object
a time series model or forecast object.

...
Additional arguments that are ignored.

Value

A numerical vector or a time series object of class ts.

Author(s)

Rob J Hyndman

[Package forecast version 5.9 Index]