Level shade bar guide

The level guide shows the colour from the forecast intervals which is blended with the series colour.

guide_level(title = waiver(), max_discrete = 5, ...)

Arguments

title
A character string or expression indicating a title of guide. If NULL, the title is not shown. By default (waiver()), the name of the scale object or the name specified in labs() is used for the title.

max_discrete
The maximum number of levels to be shown using guide_legend. If the number of levels exceeds this value, level shades are shown with guide_colourbar.

...
Further arguments passed onto either guide_colourbar or guide_legend