logLik.ets {forecast}	R Documentation

Log-Likelihood of an ets object

Description

Returns the log-likelihood of the ets model represented by object evaluated at the estimated parameters.

Usage

## S3 method for class 'ets'
logLik(object, ...)

Arguments

object
an object of class ets, representing an exponential smoothing state space model.

...
some methods for this generic require additional arguments. None are used in this method.

Value

the log-likelihood of the model represented by object evaluated at the estimated parameters.

Author(s)

Rob J Hyndman

References

Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D. (2008) Forecasting with exponential smoothing: the state space approach, Springer-Verlag. http://www.exponentialsmoothing.net.

See Also

ets

Examples

fit <- ets(USAccDeaths)
logLik(fit)

[Package forecast version 5.9 Index]