ma {forecast}	R Documentation

Moving-average smoothing

Description

ma computes a simple moving average smoother of a given time series.

Usage

ma(x, order, centre = TRUE)

Arguments

x
Univariate time series

order
Order of moving average smoother

centre
If TRUE, then the moving average is centred for even orders.

Details

The moving average smoother averages the nearest order periods of each observation. As neighbouring observations of a time series are likely to be similar in value, averaging eliminates some of the randomness in the data, leaving a smooth trend-cycle component.

T[t]=1/m(y[t-k]+y[t-k+1]+…+y[t]+…+y[t+k-1]+y[t+k])

where k=(m-1)/2

When an even order is specified, the observations averaged will include one more observation from the future than the past (k is rounded up). If centre is TRUE, the value from two moving averages (where k is rounded up and down respectively) are averaged, centering the moving average.

Value

Numerical time series object containing the simple moving average smoothed values.

Author(s)

Rob J Hyndman

See Also

decompose

Examples

plot(wineind)
sm <- ma(wineind,order=12)
lines(sm,col="red")

[Package forecast version 8.5 Index]