mstl {forecast}	R Documentation

Multiple seasonal decomposition

Description

Decompose a time series into seasonal, trend and remainder components. Seasonal components are estimated iteratively using STL. Multiple seasonal periods are allowed. The trend component is computed for the last iteration of STL. Non-seasonal time series are decomposed into trend and remainder only. In this case, supsmu is used to estimate the trend. Optionally, the time series may be Box-Cox transformed before decomposition. Unlike stl, mstl is completely automated.

Usage

mstl(x, lambda = NULL, iterate = 2, s.window = 13, ...)

Arguments

x
Univariate time series of class msts or ts.

lambda
Box-Cox transformation parameter. If lambda="auto", then a transformation is automatically selected using BoxCox.lambda. The transformation is ignored if NULL. Otherwise, data transformed before model is estimated.

iterate
Number of iterations to use to refine the seasonal component.

s.window
Seasonal windows to be used in the decompositions. If scalar, the same value is used for all seasonal components. Otherwise, it should be a vector of the same length as the number of seasonal components.

...
Other arguments are passed to stl.

See Also

stl, supsmu

Examples

library(ggplot2)
mstl(taylor) %>% autoplot()
mstl(AirPassengers, lambda = "auto") %>% autoplot()

[Package forecast version 8.13 Index]