naive {forecast}	R Documentation

Naive forecasts

Description

naive() returns forecasts and prediction intervals for an ARIMA(0,1,0) random walk model applied to x. snaive() returns forecasts and prediction intervals from an ARIMA(0,0,0)(0,1,0)m model where m is the seasonal period.

Usage

naive(x, h=10, level=c(80,95), fan=FALSE, lambda=NULL)
snaive(x, h=2*frequency(x), level=c(80,95), fan=FALSE, lambda=NULL)

Arguments

x
a numeric vector or time series

h
Number of periods for forecasting

level
Confidence levels for prediction intervals.

fan
If TRUE, level is set to seq(50,99,by=1). This is suitable for fan plots.

lambda
Box-Cox transformation parameter. Ignored if NULL. Otherwise, forecasts back-transformed via an inverse Box-Cox transformation.

Details

These functions are simply convenient wrappers to Arima with the appropriate arguments to return naive and seasonal naive forecasts.

Value

An object of class "forecast".

The function summary is used to obtain and print a summary of the results, while the function plot produces a plot of the forecasts and prediction intervals.

The generic accessor functions fitted.values and residuals extract useful features of the value returned by naive or snaive.

An object of class "forecast" is a list containing at least the following elements:

model
A list containing information about the fitted model

method
The name of the forecasting method as a character string

mean
Point forecasts as a time series

lower
Lower limits for prediction intervals

upper
Upper limits for prediction intervals

level
The confidence values associated with the prediction intervals

x
The original time series (either object itself or the time series used to create the model stored as object).

residuals
Residuals from the fitted model. That is x minus fitted values.

fitted
Fitted values (one-step forecasts)

Author(s)

Rob J Hyndman

See Also

Arima, rwf

Examples

plot(naive(gold,h=50),include=200)
plot(snaive(wineind))

[Package forecast version 5.9 Index]