ocsb.test {forecast}	R Documentation

Osborn, Chui, Smith, and Birchenhall Test for Seasonal Unit Roots

Description

An implementation of the Osborn, Chui, Smith, and Birchenhall (OCSB) test.

Usage

ocsb.test(x, lag.method = c("fixed", "AIC", "BIC", "AICc"), maxlag = 0)

Arguments

x
a univariate seasonal time series.

lag.method
a character specifying the lag order selection method.

maxlag
the maximum lag order to be considered by lag.method.

Details

The regression equation may include lags of the dependent variable. When lag.method = "fixed", the lag order is fixed to maxlag; otherwise, maxlag is the maximum number of lags considered in a lag selection procedure that minimises the lag.method criterion, which can be AIC or BIC or corrected AIC, AICc, obtained as AIC + (2k(k+1))/(n-k-1), where k is the number of parameters and n is the number of available observations in the model.

Critical values for the test are based on simulations, which has been smoothed over to produce critical values for all seasonal periods.

Value

ocsb.test returns a list of class "OCSBtest" with the following components: * statistics the value of the test statistics. * pvalues the p-values for each test statistics. * method a character string describing the type of test. * data.name a character string giving the name of the data. * fitted.model the fitted regression model.

References

Osborn DR, Chui APL, Smith J, and Birchenhall CR (1988) "Seasonality and the order of integration for consumption", Oxford Bulletin of Economics and Statistics 50(4):361-377.

See Also

nsdiffs

Examples

ocsb.test(AirPassengers)

[Package forecast version 8.10 Index]