seasonaldummy {forecast}	R Documentation

Seasonal dummy variables

Description

seasonaldummy returns a matrix of dummy variables suitable for use in Arima, auto.arima or tslm. The last season is omitted and used as the control.

Usage

seasonaldummy(x, h = NULL)

seasonaldummyf(x, h)

Arguments

x
Seasonal time series: a ts or a msts object

h
Number of periods ahead to forecast (optional)

Details

seasonaldummyf is deprecated, instead use the h argument in seasonaldummy.

The number of dummy variables is determined from the time series characteristics of x. When h is missing, the length of x also determines the number of rows for the matrix returned by seasonaldummy. the value of h determines the number of rows for the matrix returned by seasonaldummy, typically used for forecasting. The values within x are not used.

Value

Numerical matrix.

Author(s)

Rob J Hyndman

See Also

fourier

Examples

plot(ldeaths)

# Using seasonal dummy variables
month <- seasonaldummy(ldeaths)
deaths.lm  <- tslm(ldeaths ~ month)
tsdisplay(residuals(deaths.lm))
ldeaths.fcast <- forecast(deaths.lm,
    data.frame(month=I(seasonaldummy(ldeaths,36))))
plot(ldeaths.fcast)

# A simpler approach to seasonal dummy variables
deaths.lm  <- tslm(ldeaths ~ season)
ldeaths.fcast <- forecast(deaths.lm, h=36)
plot(ldeaths.fcast)

[Package forecast version 8.5 Index]