seasonplot {forecast}	R Documentation

Seasonal plot

Description

Plots a seasonal plot as described in Hyndman and Athanasopoulos (2014, chapter 2).

Usage

seasonplot(x, s, season.labels=NULL, year.labels=FALSE,
    year.labels.left=FALSE, type="o", main, ylab="",
    xlab=NULL, col=1, labelgap=0.1, ...)

Arguments

x
a numeric vector or time series.

s
seasonal frequency of x

season.labels
Labels for each season in the "year"

year.labels
Logical flag indicating whether labels for each year of data should be plotted on the right.

year.labels.left
Logical flag indicating whether labels for each year of data should be plotted on the left.

type
plot type (as for plot)

main
Main title.

ylab
Y-axis label

xlab
X-axis label

col
Colour

labelgap
Distance between year labels and plotted lines

...
additional arguments to plot.

Value

None.

Author(s)

Rob J Hyndman

References

Hyndman and Athanasopoulos (2014) Forecasting: principles and practice, OTexts: Melbourne, Australia. http://www.otexts.org/fpp/

See Also

monthplot

Examples

seasonplot(AirPassengers,col=rainbow(12),year.labels=TRUE)

[Package forecast version 5.9 Index]