ses {forecast}	R Documentation

Exponential smoothing forecasts

Description

Returns forecasts and other information for exponential smoothing forecasts applied to x.

Usage

ses(x, h=10, level=c(80,95), fan=FALSE, 
    initial=c("optimal","simple"), alpha=NULL, ...)
holt(x, h=10, damped=FALSE, level=c(80,95), fan=FALSE, 
    initial=c("optimal","simple"), exponential=FALSE,
    alpha=NULL, beta=NULL, ...)
hw(x, h=2*frequency(x), seasonal="additive", damped=FALSE, 
    level=c(80,95), fan=FALSE, initial=c("optimal","simple"), 
    exponential=FALSE, alpha=NULL, beta=NULL, gamma=NULL, ...)

Arguments

x
a numeric vector or time series

h
Number of periods for forecasting.

damped
If TRUE, use a damped trend.

seasonal
Type of seasonality in hw model. "additive" or "multiplicative"

level
Confidence level for prediction intervals.

fan
If TRUE, level is set to seq(50,99,by=1). This is suitable for fan plots.

initial
Method used for selecting initial state values. If optimal, the initial values are optimized along with the smoothing parameters using ets. If simple, the initial values are set to values obtained using simple calculations on the first few observations. See Hyndman & Athanasopoulos (2014) for details.

exponential
If TRUE, an exponential trend is fitted. Otherwise, the trend is (locally) linear.

alpha
Value of smoothing parameter for the level. If NULL, it will be estimated.

beta
Value of smoothing parameter for the trend. If NULL, it will be estimated.

gamma
Value of smoothing parameter for the seasonal component. If NULL, it will be estimated.

...
Other arguments passed to forecast.ets.

Details

ses, holt and hw are simply convenient wrapper functions for forecast(ets(...)).

Value

An object of class "forecast".

The function summary is used to obtain and print a summary of the results, while the function plot produces a plot of the forecasts and prediction intervals.

The generic accessor functions fitted.values and residuals extract useful features of the value returned by ets and associated functions.

An object of class "forecast" is a list containing at least the following elements:

model
A list containing information about the fitted model

method
The name of the forecasting method as a character string

mean
Point forecasts as a time series

lower
Lower limits for prediction intervals

upper
Upper limits for prediction intervals

level
The confidence values associated with the prediction intervals

x
The original time series (either object itself or the time series used to create the model stored as object).

residuals
Residuals from the fitted model. That is x minus fitted values.

fitted
Fitted values (one-step forecasts)

Author(s)

Rob J Hyndman

References

Hyndman, R.J., Koehler, A.B., Ord, J.K., Snyder, R.D. (2008) Forecasting with exponential smoothing: the state space approach, Springer-Verlag: New York. http://www.exponentialsmoothing.net.

Hyndman, R.J., Athanasopoulos (2014) Forecasting: principles and practice, OTexts: Melbourne, Australia. http://www.otexts.org/fpp.

See Also

ets, HoltWinters, rwf, arima.

Examples

fcast <- holt(airmiles)
plot(fcast)
deaths.fcast <- hw(USAccDeaths,h=48)
plot(deaths.fcast)

[Package forecast version 5.9 Index]