splinef {forecast}	R Documentation

Cubic Spline Forecast

Description

Returns local linear forecasts and prediction intervals using cubic smoothing splines.

Usage

splinef(y, h = 10, level = c(80, 95), fan = FALSE, lambda = NULL,
  biasadj = FALSE, method = c("gcv", "mle"), x = y)

Arguments

y
a numeric vector or time series of class ts

h
Number of periods for forecasting

level
Confidence level for prediction intervals.

fan
If TRUE, level is set to seq(51,99,by=3). This is suitable for fan plots.

lambda
Box-Cox transformation parameter. If lambda="auto", then a transformation is automatically selected using BoxCox.lambda. The transformation is ignored if NULL. Otherwise, data transformed before model is estimated.

biasadj
Use adjusted back-transformed mean for Box-Cox transformations. If transformed data is used to produce forecasts and fitted values, a regular back transformation will result in median forecasts. If biasadj is TRUE, an adjustment will be made to produce mean forecasts and fitted values.

method
Method for selecting the smoothing parameter. If method="gcv", the generalized cross-validation method from smooth.spline is used. If method="mle", the maximum likelihood method from Hyndman et al (2002) is used.

x
Deprecated. Included for backwards compatibility.

Details

The cubic smoothing spline model is equivalent to an ARIMA(0,2,2) model but with a restricted parameter space. The advantage of the spline model over the full ARIMA model is that it provides a smooth historical trend as well as a linear forecast function. Hyndman, King, Pitrun, and Billah (2002) show that the forecast performance of the method is hardly affected by the restricted parameter space.

Value

An object of class "forecast".

The function summary is used to obtain and print a summary of the results, while the function plot produces a plot of the forecasts and prediction intervals.

The generic accessor functions fitted.values and residuals extract useful features of the value returned by splinef.

An object of class "forecast" containing the following elements:

model
A list containing information about the fitted model

method
The name of the forecasting method as a character string

mean
Point forecasts as a time series

lower
Lower limits for prediction intervals

upper
Upper limits for prediction intervals

level
The confidence values associated with the prediction intervals

x
The original time series (either object itself or the time series used to create the model stored as object).

onestepf
One-step forecasts from the fitted model.

fitted
Smooth estimates of the fitted trend using all data.

residuals
Residuals from the fitted model. That is x minus one-step forecasts.

Author(s)

Rob J Hyndman

References

Hyndman, King, Pitrun and Billah (2005) Local linear forecasts using cubic smoothing splines. Australian and New Zealand Journal of Statistics, 47(1), 87-99. https://robjhyndman.com/publications/splinefcast/.

See Also

smooth.spline, arima, holt.

Examples

fcast <- splinef(uspop,h=5)
plot(fcast)
summary(fcast)

[Package forecast version 8.6 Index]