tsdisplay {forecast}	R Documentation

Time series display

Description

Plots a time series along with its acf and either its pacf, lagged scatterplot or spectrum.

Usage

tsdisplay(x, plot.type=c("partial","scatter","spectrum"), points=TRUE, ci.type="white", 
    lag.max, na.action=na.contiguous, 
    main=NULL, ylab="", xlab="", pch=1, cex=0.5, ...)

Arguments

x
a numeric vector or time series.

plot.type
type of plot to include in lower right corner.

points
logical flag indicating whether to show the individual points or not in the time plot.

ci.type
type of confidence limits for ACF. Possible values are as for acf.

lag.max
the maximum lag to plot for the acf and pacf. A suitable value is selected by default if the argument is missing.

na.action
function to handle missing values in acf, pacf and spectrum calculations. The default is na.contiguous. Useful alternatives are na.pass and na.interp.

main
Main title.

ylab
Y-axis label

xlab
X-axis label

pch
Plotting character

cex
Character size

...
additional arguments to acf.

Value

None.

Author(s)

Rob J Hyndman

References

Hyndman and Athanasopoulos (2014) Forecasting: principles and practice, OTexts: Melbourne, Australia. http://www.otexts.org/fpp/

See Also

plot.ts, Acf, spec.ar

Examples

tsdisplay(diff(WWWusage))

[Package forecast version 5.9 Index]