tsoutliers {forecast}	R Documentation

Identify and replace outliers in a time series

Description

Uses supsmu for non-seasonal series and a periodic stl decomposition with seasonal series to identify outliers and estimate their replacements.

Usage

tsoutliers(x, iterate = 2, lambda = NULL)

Arguments

x
time series

iterate
the number of iteration only for non-seasonal series

lambda
Box-Cox transformation parameter. If lambda="auto", then a transformation is automatically selected using BoxCox.lambda. The transformation is ignored if NULL. Otherwise, data transformed before model is estimated.

Value

index
Indicating the index of outlier(s)

replacement
Suggested numeric values to replace identified outliers

Author(s)

Rob J Hyndman

See Also

na.interp, tsclean

Examples

data(gold)
tsoutliers(gold)

[Package forecast version 8.6 Index]