k-fold Cross-Validation applied to an autoregressive model

CVar computes the errors obtained by applying an autoregressive modelling function to subsets of the time series y using k-fold cross-validation as described in Bergmeir, Hyndman and Koo (2015).

CVar(y, k = 10, FUN = nnetar, cvtrace = FALSE, ...)

Arguments

y
Univariate time series

k
Number of folds to use for cross-validation.

FUN
Function to fit an autoregressive model. Currently, it only works with the nnetar function.

cvtrace
Provide progress information.

...
Other arguments are passed to FUN.

Value

A list containing information about the model and accuracy for each fold, plus other summary information computed across folds.

References

Bergmeir, C., Hyndman, R.J., Koo, B. (2015) A note on the validity of cross-validation for evaluating time series prediction. Monash working paper 10/15. https://robjhyndman.com/publications/cv-time-series/.

See also

CV, tsCV.

Examples

modelcv <- CVar(lynx, k=5, lambda=0.15)
print(modelcv)

#> Series: lynx 
#> Call:   CVar(y = lynx, k = 5, lambda = 0.15)
#> 
#> 5-fold cross-validation
#>                    Mean          SD
#> ME        -117.82171245 207.4297895
#> RMSE       865.97203798 245.8837532
#> MAE        578.72048791 190.6054503
#> MPE        -27.84715313  35.0711508
#> MAPE        60.02540438  27.5811199
#> ACF1        -0.01593769   0.2704894
#> Theil's U    0.90977105   0.1692362

print(modelcv$fold1)

#> $model
#> Series: y 
#> Model:  NNAR(11,6) 
#> Call:   FUN(y = y, lambda = 0.15, subset = trainset)
#> 
#> Average of 20 networks, each of which is
#> a 11-6-1 network with 79 weights
#> options were - linear output units 
#> 
#> sigma^2 estimated as 0.0949
#> 
#> $accuracy
#>                 ME     RMSE      MAE       MPE     MAPE      ACF1 Theil's U
#> Test set -352.3341 1070.599 760.3141 -89.87951 109.2968 0.3525681 0.9065455
#>