arima.errors {forecast}	R Documentation

Errors from a regression model with ARIMA errors

Description

Returns time series of the regression residuals from a fitted ARIMA model.

Usage

arima.errors(object)

Arguments

object
An object containing a time series model of class Arima.

Details

This is a deprecated function which is identical to residuals.Arima(object, type="regression") Regression residuals are equal to the original data minus the effect of any regression variables. If there are no regression variables, the errors will be identical to the original series (possibly adjusted to have zero mean).

Value

A ts object

Author(s)

Rob J Hyndman

See Also

residuals.Arima.

[Package forecast version 8.5 Index]