Return the order of an ARIMA or ARFIMA model

Returns the order of a univariate ARIMA or ARFIMA model.

arimaorder(object)

Arguments

object
An object of class “Arima”, dQuotear or “fracdiff”. Usually the result of a call to arima, Arima, auto.arima, ar, arfima or fracdiff.

Value

A numerical vector giving the values \(p\), \(d\) and \(q\) of the ARIMA or ARFIMA model. For a seasonal ARIMA model, the returned vector contains the values \(p\), \(d\), \(q\), \(P\), \(D\), \(Q\) and \(m\), where \(m\) is the period of seasonality.

See also

Author

Rob J Hyndman

Examples

WWWusage %>% auto.arima %>% arimaorder

#> p d q 
#> 1 1 1