autoplot.decomposed.ts {forecast}	R Documentation

Plot time series decomposition components using ggplot

Description

Produces a ggplot object of seasonally decomposed time series for objects of class “stl” (created with stl), class “seas” (created with seas), or class “decomposed.ts” (created with decompose).

Usage

## S3 method for class 'decomposed.ts'
autoplot(object, labels = NULL, range.bars = NULL, ...)

## S3 method for class 'stl'
autoplot(object, labels = NULL, range.bars = TRUE, ...)

## S3 method for class 'StructTS'
autoplot(object, labels = NULL, range.bars = TRUE, ...)

## S3 method for class 'seas'
autoplot(object, labels = NULL, range.bars = NULL, ...)

## S3 method for class 'mstl'
autoplot(object, ...)

Arguments

object
Object of class “seas”, “stl”, or “decomposed.ts”.

labels
Labels to replace “seasonal”, “trend”, and “remainder”.

range.bars
Logical indicating if each plot should have a bar at its right side representing relative size. If NULL, automatic selection takes place.

...
Other plotting parameters to affect the plot.

Value

Returns an object of class ggplot.

Author(s)

Mitchell O'Hara-Wild

See Also

seas, stl, decompose, StructTS, plot.stl.

Examples

library(ggplot2)
co2 %>% decompose %>% autoplot
nottem %>% stl(s.window='periodic') %>% autoplot

## Not run: 
library(seasonal)
seas(USAccDeaths) %>% autoplot

## End(Not run)

[Package forecast version 8.10 Index]