bizdays {forecast}	R Documentation

Number of trading days in each season

Description

Returns number of trading days in each month or quarter of the observed time period.

Usage

bizdays(x, FinCenter)

Arguments

x
Monthly or quarterly time series

FinCenter
A character with the the location of the financial center named as "continent/city". This concept allows to handle data records collected in different time zones and mix them up to have always the proper time stamps with respect to your personal financial center, or alternatively to the GMT reference time. More details on finCenter.

Details

Useful for trading days length adjustments. More on how to define "business days", please refer to isBizday.

Value

Time series

Author(s)

Earo Wang

See Also

monthdays

Examples

  bizdays(wineind, FinCenter = "Sydney")

[Package forecast version 5.9 Index]