bizdays {forecast}	R Documentation

Number of trading days in each season

Description

Returns number of trading days in each month or quarter of the observed time period in a major financial center.

Usage

bizdays(x, FinCenter = c("New York", "London", "NERC", "Tokyo",
  "Zurich"))

Arguments

x
Monthly or quarterly time series

FinCenter
Major financial center.

Details

Useful for trading days length adjustments. More on how to define "business days", please refer to isBizday.

Value

Time series

Author(s)

Earo Wang

See Also

monthdays

Examples

  x <-  ts(rnorm(30), start = c(2013, 2), frequency = 12)
  bizdays(x, FinCenter = "New York")

[Package forecast version 8.6 Index]