forecast: Forecasting Functions for Time Series and Linear Models

Methods and tools for displaying and analysing univariate time series forecasts including exponential smoothing via state space models and automatic ARIMA modelling.

See also

Author

Maintainer: Rob Hyndman Rob.Hyndman@monash.edu (ORCID) [copyright holder]

Authors:

George Athanasopoulos (ORCID)

Christoph Bergmeir (ORCID)

Gabriel Caceres (ORCID)

Leanne Chhay

Kirill Kuroptev

Maximilian Mücke (ORCID)

Mitchell O'Hara-Wild (ORCID)

Fotios Petropoulos (ORCID)

Slava Razbash

Earo Wang (ORCID)

Farah Yasmeen (ORCID)

Other contributors:

Federico Garza [contributor]

Daniele Girolimetto [contributor]

Ross Ihaka [contributor, copyright holder]

R Core Team [contributor, copyright holder]

Daniel Reid [contributor]

David Shaub [contributor]

Yuan Tang (ORCID) [contributor]

Xiaoqian Wang [contributor]

Zhenyu Zhou [contributor]