Forecasting using BATS and TBATS models

Forecasts h steps ahead with a BATS model. Prediction intervals are also produced.

# S3 method for bats
forecast(object, h, level = c(80, 95), fan = FALSE, biasadj = NULL, ...)

# S3 method for tbats
forecast(object, h, level = c(80, 95), fan = FALSE, biasadj = NULL, ...)

Arguments

object
An object of class "bats". Usually the result of a call to bats.

h
Number of periods for forecasting. Default value is twice the largest seasonal period (for seasonal data) or ten (for non-seasonal data).

level
Confidence level for prediction intervals.

fan
If TRUE, level is set to seq(51,99,by=3). This is suitable for fan plots.

biasadj
Use adjusted back-transformed mean for Box-Cox transformations. If TRUE, point forecasts and fitted values are mean forecast. Otherwise, these points can be considered the median of the forecast densities.

...
Other arguments, currently ignored.

Value

An object of class "forecast".

The function summary is used to obtain and print a summary of the results, while the function plot produces a plot of the forecasts and prediction intervals.

The generic accessor functions fitted.values and residuals extract useful features of the value returned by forecast.bats.

An object of class "forecast" is a list containing at least the following elements:

model

A copy of the bats object

method

The name of the forecasting method as a character string

mean

Point forecasts as a time series

lower

Lower limits for prediction intervals

upper

Upper limits for prediction intervals

level

The confidence values associated with the prediction intervals

x

The original time series (either object itself or the time series used to create the model stored as object).

residuals

Residuals from the fitted model.

fitted

Fitted values (one-step forecasts)

References

De Livera, A.M., Hyndman, R.J., & Snyder, R. D. (2011), Forecasting time series with complex seasonal patterns using exponential smoothing, Journal of the American Statistical Association, 106(496), 1513-1527.

See also

Author

Slava Razbash and Rob J Hyndman

Examples

if (FALSE) {
fit <- bats(USAccDeaths)
plot(forecast(fit))

taylor.fit <- bats(taylor)
plot(forecast(taylor.fit))
}