Get response variable from time series model.

getResponse is a generic function for extracting the historical data from a time series model (including Arima, ets, ar, fracdiff), a linear model of class lm, or a forecast object. The function invokes particular methods which depend on the class of the first argument.

getResponse(object, ...)

# S3 method for default
getResponse(object, ...)

# S3 method for lm
getResponse(object, ...)

# S3 method for Arima
getResponse(object, ...)

# S3 method for fracdiff
getResponse(object, ...)

# S3 method for ar
getResponse(object, ...)

# S3 method for tbats
getResponse(object, ...)

# S3 method for bats
getResponse(object, ...)

# S3 method for mforecast
getResponse(object, ...)

# S3 method for baggedModel
getResponse(object, ...)

Arguments

object
a time series model or forecast object.

...
Additional arguments that are ignored.

Value

A numerical vector or a time series object of class ts.